+++ tests/ge/taxistgela/dao/DaoTests.java
@@ -6,7 +6,6 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import static org.junit.Assert.*;
@@ -238,7 +237,7 @@ public class DaoTests {
         pref.setMinimumUserRating(0.0);//gela bizatkaznia
         man.insertDriverPreference(pref);
 
-        Location l = new Location(new BigDecimal(2.234),new BigDecimal(3.1245));
+        Location l = new Location(new Double(2.234), new Double(3.1245));
         Driver driver = new Driver(-1,"01010101011","gela@taxistgela.ge","Madridista1",1,"gela","magaltadze",Gender.MALE,"555696996",car,"gelandara95","bozandara",l,2.2,pref,true,true);
 
         man.registerDriver(driver);
